// frontend/lib/api/generated/model/salesActivityCreateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { SalesActivityCreateRequestActivityPurpose } from './salesActivityCreateRequestActivityPurpose';
import type { SalesActivityCreateRequestActivityType } from './salesActivityCreateRequestActivityType';
import type { SalesActivityCreateRequestStatus } from './salesActivityCreateRequestStatus';

export interface SalesActivityCreateRequest {
  projectOpportunityId?: number;
  activityType?: SalesActivityCreateRequestActivityType;
  activityPurpose?: SalesActivityCreateRequestActivityPurpose;
  activityContent?: string;
  location?: string;
  activityDateTime?: string;
  issue?: string;
  nextActivity?: string;
  attendeeUserIds?: string[];
  customerInterest?: string;
  status?: SalesActivityCreateRequestStatus;
  salesActivityRequestId?: number;
}