// frontend/lib/api/generated/model/salesActivityCreateRequestActivityType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */

export type SalesActivityCreateRequestActivityType = typeof SalesActivityCreateRequestActivityType[keyof typeof SalesActivityCreateRequestActivityType];


export const SalesActivityCreateRequestActivityType = {
  EMAIL: 'EMAIL',
  CALL: 'CALL',
  VIDEO_MEETING: 'VIDEO_MEETING',
  OFFLINE_MEETING: 'OFFLINE_MEETING',
  ETC: 'ETC',
} as const;