// frontend/lib/api/generated/model/salesActivityCreateRequestStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */

export type SalesActivityCreateRequestStatus = typeof SalesActivityCreateRequestStatus[keyof typeof SalesActivityCreateRequestStatus];


export const SalesActivityCreateRequestStatus = {
  REQUESTED: 'REQUESTED',
  PLANNED: 'PLANNED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELED: 'CANCELED',
} as const;