// frontend/lib/api/generated/model/salesActivityResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { SalesActivityResponseActivityPurpose } from './salesActivityResponseActivityPurpose';
import type { SalesActivityResponseActivityType } from './salesActivityResponseActivityType';
import type { SalesActivityResponseStatus } from './salesActivityResponseStatus';

export interface SalesActivityResponse {
  id?: number;
  projectOpportunityId?: number;
  activityType?: SalesActivityResponseActivityType;
  activityPurpose?: SalesActivityResponseActivityPurpose;
  activityContent?: string;
  location?: string;
  activityDateTime?: string;
  issue?: string;
  nextActivity?: string;
  customerInterest?: string;
  status?: SalesActivityResponseStatus;
  salesActivityRequestId?: number;
}