// frontend/lib/api/generated/model/salesActivityUpdateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { SalesActivityUpdateRequestActivityPurpose } from './salesActivityUpdateRequestActivityPurpose';
import type { SalesActivityUpdateRequestActivityType } from './salesActivityUpdateRequestActivityType';
import type { SalesActivityUpdateRequestStatus } from './salesActivityUpdateRequestStatus';

export interface SalesActivityUpdateRequest {
  activityType?: SalesActivityUpdateRequestActivityType;
  activityPurpose?: SalesActivityUpdateRequestActivityPurpose;
  activityContent?: string;
  location?: string;
  activityDateTime?: string;
  issue?: string;
  nextActivity?: string;
  attendeeUserIds?: string[];
  customerInterest?: string;
  status?: SalesActivityUpdateRequestStatus;
}